LayoutTests/imported/w3c/web-platform-tests/websockets/security/002-expected.txt

CONSOLE MESSAGE: line 17: Refused to set unsafe header "Sec-WebSocket-Key"

PASS WebSockets: check Sec-WebSocket-Key 
